All generated files in gen folder of source
html and xml files generaed must go in the gen folder. Also, this folder must be
in the same folder as the source. This was not the case when the scripts were
added to crontab (the html and xml were generated in the home folder then.)
Also, make the python scripts executable so as to make them usable for crontabs.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>