Finally, schedule to run using the database
The local json files setting was not recommended and did not work on a ephemeral
file storage system or where read write permissions are not granted.
To get rid of that, support for using a postgres database is added.

Simply, export DATABASE_URL as an environmental variable. Then use _db() methods
to use the database.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>