Automatic cookie handling using requests Sessions Closes #14
If interacting with the same server over multiple requests, it is advisable to
use Sessions object. It improves efficiency of the code.

Also, manual handling of cookies is eliminated. This leads to concise, more
understandable code.

Also removed property setters in the TpoSession object. The reasons to have it
were not justifiable. It was ugly as well :)

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>